requirements/dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/raw/dev.txt -o requirements/dev.txt --allow-unsafe --strip-extras --no-strip-markers
-e ./benchmarks
    # via -r requirements/raw/bench.txt
-e ./tests/tests_helpers
    # via
    #   -r requirements/raw/_test_base.txt
    #   -r requirements/raw/bench.txt
aiosqlite==0.21.0
    # via -r requirements/raw/test_extra_new.txt
alabaster==1.0.0
    # via sphinx
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
astpath==0.9.1
    # via -r requirements/raw/lint.txt
attrs==24.2.0
    # via
    #   -r requirements/raw/test_extra_new.txt
    #   cattrs
    #   jsonschema
    #   referencing
babel==2.17.0
    # via sphinx
beautifulsoup4==4.13.4
    # via furo
cachetools==5.5.2
    # via tox
cattrs==23.1.2
    # via -r requirements/raw/bench.txt
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via tox
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via towncrier
colorama==0.4.6
    # via
    #   radon
    #   tox
coverage==7.8.1
    # via
    #   -r requirements/raw/_test_base.txt
    #   -r requirements/raw/runner.txt
dataclass-factory==2.16
    # via -r requirements/raw/bench.txt
dirty-equals==0.9.0
    # via -r requirements/raw/_test_base.txt
distlib==0.3.9
    # via virtualenv
dlint==0.14.1
    # via -r requirements/raw/lint.txt
docutils==0.21.2
    # via
    #   docutils-stubs
    #   myst-parser
    #   sphinx
    #   sphinx-better-subsection
    #   sphinx-paramlinks
docutils-stubs==0.0.22
    # via -r requirements/raw/doc.txt
fastapi==0.119.1
    # via -r requirements/raw/test_extra_new.txt
filelock==3.18.0
    # via
    #   tox
    #   virtualenv
flake8==7.2.0
    # via dlint
furo==2024.8.6
    # via -r requirements/raw/doc.txt
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via -r requirements/raw/doc.txt
greenlet==3.3.1 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements/raw/test_extra_new.txt
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
imagesize==1.4.1
    # via sphinx
iniconfig==2.1.0
    # via pytest
invoke==2.2.0
    # via -r requirements/raw/runner.txt
jinja2==3.1.6
    # via
    #   myst-parser
    #   sphinx
    #   towncrier
jsonschema==4.26.0
    # via -r requirements/raw/test_extra_new.txt
jsonschema-rs==0.40.0 ; implementation_name != 'pypy'
    # via -r requirements/raw/_test_base.txt
jsonschema-specifications==2025.4.1
    # via jsonschema
librt==0.7.8 ; platform_python_implementation != 'PyPy'
    # via mypy
lxml==5.4.0
    # via astpath
mando==0.7.1
    # via radon
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   myst-parser
markupsafe==3.0.2
    # via jinja2
marshmallow==3.20.1
    # via -r requirements/raw/bench.txt
mashumaro==3.17
    # via -r requirements/raw/bench.txt
mccabe==0.7.0
    # via flake8
mdit-py-plugins==0.4.2
    # via myst-parser
mdurl==0.1.2
    # via markdown-it-py
msgspec==0.20.0 ; implementation_name != 'pypy'
    # via
    #   -r requirements/raw/bench.txt
    #   -r requirements/raw/test_extra_new.txt
mypy==1.19.1
    # via -r requirements/raw/lint.txt
mypy-extensions==1.1.0
    # via mypy
myst-parser==4.0.1
    # via -r requirements/raw/doc.txt
narwhals==1.40.0
    # via plotly
orjson==3.10.18
    # via mypy
packaging==25.0
    # via
    #   marshmallow
    #   plotly
    #   pyproject-api
    #   pytest
    #   sphinx
    #   tox
    #   tox-uv
pathspec==1.0.4
    # via mypy
pbr==6.1.1
    # via sphinxcontrib-apidoc
phonenumberslite==9.0.5
    # via -r requirements/raw/_test_base.txt
pip==25.1.1
    # via -r requirements/raw/pre.txt
platformdirs==4.3.8
    # via
    #   tox
    #   virtualenv
plotly==6.1.1
    # via
    #   -r requirements/raw/_bench-orchestrate.txt
    #   -r requirements/raw/doc.txt
pluggy==1.6.0
    # via
    #   pytest
    #   tox
prek==0.2.5
    # via -r requirements/raw/lint.txt
psutil==5.9.5
    # via
    #   -r requirements/raw/bench.txt
    #   pyperf
pycodestyle==2.13.0
    # via flake8
pydantic==2.12.5
    # via
    #   -r requirements/raw/bench.txt
    #   -r requirements/raw/test_extra_new.txt
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
pyflakes==3.3.2
    # via flake8
pygments==2.19.1
    # via
    #   furo
    #   sphinx
pyperf==2.6.1
    # via -r requirements/raw/bench.txt
pyproject-api==1.9.1
    # via tox
pytest==8.3.5
    # via
    #   -r requirements/raw/_test_base.txt
    #   -r requirements/raw/bench.txt
    #   pytest-asyncio
pytest-asyncio==1.2.0
    # via -r requirements/raw/test_extra_new.txt
pyyaml==6.0.2
    # via myst-parser
radon==6.0.1
    # via -r requirements/raw/lint.txt
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via sphinx
roman-numerals-py==3.1.0
    # via sphinx
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
ruff==0.14.14
    # via -r requirements/raw/lint.txt
schematics==2.1.1
    # via -r requirements/raw/bench.txt
setuptools==80.8.0
    # via
    #   -r requirements/raw/dev.txt
    #   pbr
six==1.17.0
    # via mando
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
snowballstemmer==3.0.1
    # via sphinx
soupsieve==2.7
    # via beautifulsoup4
sphinx==8.2.3
    # via
    #   -r requirements/raw/doc.txt
    #   furo
    #   myst-parser
    #   sphinx-basic-ng
    #   sphinx-copybutton
    #   sphinx-design
    #   sphinx-paramlinks
    #   sphinx-reredirects
    #   sphinxcontrib-apidoc
    #   sphinxcontrib-globalsubs
    #   sphinxext-opengraph
sphinx-basic-ng==1.0.0b2
    # via furo
sphinx-better-subsection==0.2
    # via -r requirements/raw/doc.txt
sphinx-copybutton==0.5.2
    # via -r requirements/raw/doc.txt
sphinx-design==0.6.1
    # via -r requirements/raw/doc.txt
sphinx-paramlinks==0.6.0
    # via -r requirements/raw/doc.txt
sphinx-reredirects==0.1.6
    # via -r requirements/raw/doc.txt
sphinxcontrib-apidoc==0.6.0
    # via -r requirements/raw/doc.txt
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-globalsubs==0.1.2
    # via -r requirements/raw/doc.txt
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
sphinxext-opengraph==0.10.0
    # via -r requirements/raw/doc.txt
sqlalchemy==2.0.46
    # via -r requirements/raw/test_extra_new.txt
starlette==0.48.0
    # via fastapi
toml==0.10.2
    # via vulture
towncrier==24.8.0
    # via -r requirements/raw/dev.txt
tox==4.26.0
    # via tox-uv
tox-uv==1.25.0
    # via -r requirements/raw/runner.txt
typing-extensions==4.15.0
    # via
    #   aiosqlite
    #   beautifulsoup4
    #   fastapi
    #   mashumaro
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
typos==1.42.3
    # via -r requirements/raw/lint.txt
urllib3==2.4.0
    # via requests
uv==0.7.6
    # via
    #   -r requirements/raw/pre.txt
    #   tox-uv
virtualenv==20.31.2
    # via tox
vulture==2.10
    # via -r requirements/raw/lint.txt
wheel==0.45.1
    # via -r requirements/raw/pre.txt
